finished task feature

+++ src/task/task.service.ts
@@ -1,4 +1,4 @@
-import { BadRequestException, Injectable } from '@nestjs/common';
+import { BadRequestException, ForbiddenException, Injectable, NotFoundException } from '@nestjs/common';
 import { CreateTaskDto } from './dto/create-task.dto';
 import { UpdateTaskDto } from './dto/update-task.dto';
 import { PrismaService } from '../prisma.service';
@@ -31,9 +31,13 @@ export class TaskService {
     } catch (error) { throw error }
   }
 
-  async findAll(): Promise<Task[]> {
+  async findAll(creator_id: number): Promise<Task[]> {
     try {
-      return await this.prisma.task.findMany({})
+      return await this.prisma.task.findMany({
+        where: {
+          creator_id: creator_id
+        }
+      })
     } catch (error) {
       throw error
     }
@@ -47,8 +51,8 @@ export class TaskService {
         }
       })
 
-      if (!task){
-        throw new BadRequestException("task doesn't exist")
+      if (!task) {
+        throw new NotFoundException("task doesn't exist")
       }
       return task
     } catch (error) {
@@ -58,14 +62,14 @@ export class TaskService {
 
   async update(id: number, updateTaskDto: UpdateTaskDto): Promise<Task | null> {
     try {
-      const taskToBeUpdated =  await this.prisma.task.findFirst({
-        where:{
+      const taskToBeUpdated = await this.prisma.task.findFirst({
+        where: {
           id: id
         }
       });
 
-      if (!taskToBeUpdated){
-        throw new BadRequestException("Task doesn't exist")
+      if (!taskToBeUpdated) {
+        throw new NotFoundException("Task doesn't exist")
       }
       const updatedTask = await this.prisma.task.update({
         where: {
@@ -79,15 +83,30 @@ export class TaskService {
     } catch (error) { throw error }
   }
 
-  async remove(id: number): Promise<void> {
+  async remove(id: number, roles: string[], user_id: number): Promise<void> {
     try {
+      const task = await this.prisma.task.findUnique({
+        where: { id: id }
+      });
+
+      if (!task) {
+        throw new NotFoundException('Task not found');
+      }
+
+      if (roles.includes('admin')) {
+        await this.prisma.task.delete({
+          where: { id: id }
+        });
+        return;
+      }
+      if (task.creator_id !== user_id) {
+        throw new ForbiddenException('You are not allowed to delete this task');
+      }
       await this.prisma.task.delete({
-        where: {
-          id: id
-        }
-      })
+        where: { id: id }
+      });
     } catch (error) {
-      throw error
+      throw error;
     }
   }
 }
